refactor: extract each helper function in its own file
