+++ packages/vee-validate/src/index.ts
@@ -7,10 +7,13 @@ export { ErrorMessage } from './ErrorMessage';
 export { useField } from './useField';
 export { useForm } from './useForm';
 export { ValidationResult, FormActions, FormState, FormValidationResult, FormContext } from './types';
-export * from './useErrors';
 export { useResetForm } from './useResetForm';
-export * from './useIsDirty';
-export * from './useIsTouched';
-export * from './useIsValid';
-export * from './useIsSubmitting';
-export * from './useValidate';
+export { useIsFieldDirty } from './useIsFieldDirty';
+export { useIsFieldTouched } from './useIsFieldTouched';
+export { useIsFieldValid } from './useIsFieldValid';
+export { useIsSubmitting } from './useIsSubmitting';
+export { useValidateField } from './useValidateField';
+export { useIsFormDirty } from './useIsFormDirty';
+export { useIsFormTouched } from './useIsFormTouched';
+export { useIsFormValid } from './useIsFormValid';
+export { useValidateForm } from './useValidateForm';
