+++ packages/vee-validate/src/useFormErrors.ts
@@ -1,0 +1,11 @@
+import { FormErrorsSymbol } from './symbols';
+import { injectWithSelf } from './utils';
+
+/**
+ * Gives access to all form errors
+ */
+export function useFormErrors() {
+  const errors = injectWithSelf(FormErrorsSymbol);
+
+  return errors;
+}
