+++ packages/vee-validate/src/useIsFieldTouched.ts
@@ -1,0 +1,14 @@
+import { computed } from 'vue';
+import { FormSymbol } from './symbols';
+import { injectWithSelf } from './utils';
+
+/**
+ * If a field is touched or not
+ */
+export function useIsFieldTouched(path: string) {
+  const form = injectWithSelf(FormSymbol);
+
+  return computed(() => {
+    return form?.fields.value[path]?.meta.touched;
+  });
+}
