+++ packages/vee-validate/src/useIsFormTouched.ts
@@ -3,17 +3,6 @@ import { FormSymbol } from './symbols';
 import { injectWithSelf } from './utils';
 
 /**
- * If a field is touched or not
- */
-export function useIsFieldTouched(path: string) {
-  const form = injectWithSelf(FormSymbol);
-
-  return computed(() => {
-    return form?.fields.value[path]?.meta.touched;
-  });
-}
-
-/**
  * If the form is touched or not
  */
 export function useIsFormTouched() {
