+++ packages/vee-validate/tests/useIsFieldDirty.spec.ts
@@ -1,5 +1,5 @@
 import flushPromises from 'flush-promises';
-import { useField, useForm, useIsFormDirty, useIsFieldDirty } from '@/vee-validate';
+import { useField, useForm, useIsFieldDirty } from '@/vee-validate';
 import { mountWithHoc, setValue } from './helpers';
 
 describe('useIsFieldDirty()', () => {
@@ -22,36 +22,6 @@ describe('useIsFieldDirty()', () => {
     `,
     });
     await flushPromises();
-
-    const input = document.querySelector('input');
-    const error = document.querySelector('span');
-    expect(error?.textContent).toBe('false');
-
-    setValue(input as any, '');
-    await flushPromises();
-    expect(error?.textContent).toBe('true');
-  });
-});
-
-describe('useIsFormDirty()', () => {
-  test('gives access to the forms isDirty status', async () => {
-    mountWithHoc({
-      setup() {
-        useForm();
-        const { value, handleInput } = useField('test');
-        const isDirty = useIsFormDirty();
-
-        return {
-          value,
-          isDirty,
-          handleInput,
-        };
-      },
-      template: `
-      <input name="field" v-model="value" @input="handleInput" />
-      <span>{{ isDirty.toString()  }}</span>
-    `,
-    });
 
     const input = document.querySelector('input');
     const error = document.querySelector('span');
