+++ packages/vee-validate/tests/useValidateForm.spec.ts
@@ -1,40 +1,11 @@
 import flushPromises from 'flush-promises';
-import { useField, useForm, useValidateField, useValidateForm } from '@/vee-validate';
+import { useField, useForm, useValidateForm } from '@/vee-validate';
 import { mountWithHoc } from './helpers';
 
-const REQUIRED_MESSAGE = 'Field is required';
-const rules = (val: any) => (val ? true : REQUIRED_MESSAGE);
-
-describe('useValidateField()', () => {
-  test('validates a single field', async () => {
-    let validate!: ReturnType<typeof useValidateField>;
-    mountWithHoc({
-      setup() {
-        useForm();
-        const { errorMessage } = useField('test', rules);
-        validate = useValidateField('test');
-
-        return {
-          errorMessage,
-        };
-      },
-      template: `
-      <span>{{ errorMessage }}</span>
-    `,
-    });
-
-    await flushPromises();
-    const error = document.querySelector('span');
-    expect(error?.textContent).toBe('');
-    await validate();
-    await flushPromises();
-
-    await flushPromises();
-    expect(error?.textContent).toBe(REQUIRED_MESSAGE);
-  });
-});
-
 describe('useValidateForm()', () => {
+  const REQUIRED_MESSAGE = 'Field is required';
+  const rules = (val: any) => (val ? true : REQUIRED_MESSAGE);
+
   test('validates all fields', async () => {
     let validate!: ReturnType<typeof useValidateForm>;
     mountWithHoc({
